Refatoração: expansão dos modelos, testes e infraestrutura de integração da API v2
- Refatoradas as APIs Pessoas e Notas Fiscais para a API v2 da ContaAzul, com modelos novos/atualizados e respostas paginadas.
- Adicionados métodos CRUD, em lote e de detalhes para as APIs PessoasApi e NotasFiscaisApi.
- Introduzidas respostas paginadas genéricas, vinculação MDF-e e recuperação de XML de faturas.
- Conjunto de testes reformulado: testes de contrato HTTP, serialização de ida e volta e testes de integração em tempo real com configuração flexível.
- Adicionada a CLI TokenHelper para aquisição de tokens OAuth.
- Documentação, exemplos, .gitignore e arquivos de projeto atualizados.
- Modelos obsoletos removidos e problemas de codificação corrigidos.

+++ contaazul-sdk-net/Models/Vendas/VendaQuantidades.cs
@@ -9,18 +9,18 @@ namespace ContaAzul.Sdk.Net.Models.Vendas
     {
         /// <summary>Quantidade total de vendas.</summary>
         [JsonPropertyName("total")]
-        public int Total { get; set; }
+        public decimal Total { get; set; }
 
         /// <summary>Quantidade de vendas aprovadas.</summary>
         [JsonPropertyName("aprovado")]
-        public int Aprovado { get; set; }
+        public decimal Aprovado { get; set; }
 
         /// <summary>Quantidade de vendas canceladas.</summary>
         [JsonPropertyName("cancelado")]
-        public int Cancelado { get; set; }
+        public decimal Cancelado { get; set; }
 
         /// <summary>Quantidade de vendas esperando aprovação.</summary>
         [JsonPropertyName("esperando_aprovacao")]
-        public int EsperandoAprovacao { get; set; }
+        public decimal EsperandoAprovacao { get; set; }
     }
 }
